# openapi-spec.yaml
openapi: 3.0.3
info:
  title: Multi-Platform PDF Services API
  description: |
    Kompleksowy zestaw serwisów PDF implementowanych w różnych technologiach.
    Wszystkie serwisy mają identyczne API dla łatwego porównywania i testowania.
    
    ## Dostępne Serwisy:
    - **IronPDFService** (.NET + IronPDF) - Port 5001
    - **PdfSharpService** (.NET + PdfSharpCore) - Port 5002  
    - **PythonPdfService** (Python + Flask) - Port 5003
    - **NodeJSPdfService** (Node.js + Express) - Port 5033
    - **GoPdfService** (Go + Gin) - Port 5034
    
    ## Funkcjonalności:
    - ✅ Łączenie plików PDF
    - ✅ Konwersja obrazów do PDF
    - ✅ Łączenie PDF i obrazów
    - ✅ Ekstrakcja tekstu z PDF
    - ✅ Konwersja DOC/DOCX do PDF
    - ✅ Dodawanie kodów QR i kreskowych
    - ✅ Podpisy cyfrowe i weryfikacja
    - ✅ Odczytywanie kodów kreskowych i QR
  version: 1.0.0
  contact:
    name: PDF Services Team
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001
    description: IronPDFService (.NET + IronPDF)
  - url: http://localhost:5002
    description: PdfSharpService (.NET + PdfSharpCore)
  - url: http://localhost:5003
    description: PythonPdfService (Python + Flask)
  - url: http://localhost:5033
    description: NodeJSPdfService (Node.js + Express)
  - url: http://localhost:5034
    description: GoPdfService (Go + Gin)
...
paths:
...
  /api/pdf/merge-pdfs:
    post:
      tags:
        - PDF Operations
      summary: Łączy kilka plików PDF w jeden dokument
      description: Łączy przekazane pliki PDF w jeden dokument w określonym formacie
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Pliki PDF do połączenia
                outputFormat:
                  type: string
                  enum: [A4, A3, A5, LETTER]
                  default: A4
                  description: Format wyjściowy dokumentu
      responses:
        '200':
          description: Pliki PDF zostały połączone
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Błąd w żądaniu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Błąd serwera
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/pdf/images-to-pdf:
    post:
      tags:
        - PDF Operations
      summary: Konwertuje obrazy do formatu PDF
      description: Konwertuje przekazane obrazy do formatu PDF w określonym formacie
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Pliki obrazów do konwersji
                outputFormat:
                  type: string
                  enum: [A4, A3, A5, LETTER]
                  default: A4
                  description: Format wyjściowy dokumentu
      responses:
        '200':
          description: Obrazy zostały przekonwertowane do PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Błąd w żądaniu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Błąd serwera
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/pdf/merge-all:
    post:
      tags:
        - PDF Operations
      summary: Łączy pliki PDF i obrazy w jeden dokument
      description: Łączy przekazane pliki PDF i obrazy w jeden dokument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Pliki PDF i obrazy do połączenia
                outputFormat:
                  type: string
                  enum: [A4, A3, A5, LETTER]
                  default: A4
                  description: Format wyjściowy dokumentu
      responses:
        '200':
          description: Pliki zostały połączone
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Błąd w żądaniu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Błąd serwera
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/pdf/extract-text:
    post:
      tags:
        - PDF Operations
      summary: Ekstrahuje tekst z pliku PDF
      description: Wyodrębnia tekst z przekazanego pliku PDF
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Plik PDF do ekstrakcji tekstu
      responses:
        '200':
          description: Tekst został wyekstrahowany
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Błąd w żądaniu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Błąd serwera
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/pdf/read-barcodes:
    post:
      tags:
        - Code Reading
      summary: Odczytywanie kodów kreskowych z obrazu/PDF
      description: Odczytywanie kodów kreskowych z przekazanego pliku
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Plik obrazu lub PDF
      responses:
        '200':
          description: Kody kreskowe zostały odczytane
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BarcodeReadingResponse'
        '400':
          description: Błąd w żądaniu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Błąd serwera
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/pdf/read-qr-codes:
    post:
      tags:
        - Code Reading
      summary: Odczytywanie kodów QR z obrazu/PDF
      description: Odczytywanie kodów QR z przekazanego pliku
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Plik obrazu lub PDF
      responses:
        '200':
          description: Kody QR zostały odczytane
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QRCodeReadingResponse'
        '400':
          description: Błąd w żądaniu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Błąd serwera
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/pdf/read-all-codes:
    post:
      tags:
        - Code Reading
      summary: Odczytywanie wszystkich kodów (kreskowych i QR) z obrazu/PDF
      description: Odczytywanie wszystkich kodów z przekazanego pliku
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Plik obrazu lub PDF
      responses:
        '200':
          description: Wszystkie kody zostały odczytane
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllCodesReadingResponse'
        '400':
          description: Błąd w żądaniu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Błąd serwera
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/pdf/supported-formats:
    get:
      tags:
        - Information
      summary: Zwraca informacje o obsługiwanych formatach
      description: Zwraca listę obsługiwanych formatów plików i funkcji
      responses:
        '200':
          description: Informacje o obsługiwanych formatach
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportedFormatsResponse'
        '500':
          description: Błąd serwera
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Opis błędu
          example: "Nie przekazano pliku"
    BarcodeReadingResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        barcodes:
          type: array
          items:
            $ref: '#/components/schemas/BarcodeInfo'
        count:
          type: integer
          example: 2
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    QRCodeReadingResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        qrCodes:
          type: array
          items:
            $ref: '#/components/schemas/QRCodeInfo'
        count:
          type: integer
          example: 1
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    AllCodesReadingResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        codes:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/BarcodeInfo'
              - $ref: '#/components/schemas/QRCodeInfo'
        barcodes:
          type: array
          items:
            $ref: '#/components/schemas/BarcodeInfo'
        qrCodes:
          type: array
          items:
            $ref: '#/components/schemas/QRCodeInfo'
        totalCount:
          type: integer
          example: 3
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    BarcodeInfo:
      type: object
      properties:
        type:
          type: string
          example: "barcode"
        data:
          type: string
          example: "123456789"
        format:
          type: string
          example: "CODE_128"
        confidence:
          type: number
          format: float
          example: 1.0
        page:
          type: integer
          description: Numer strony (dla PDF)
          example: 1
        bounds:
          $ref: '#/components/schemas/Bounds'
...
    Bounds:
      type: object
      properties:
        x:
          type: integer
          example: 100
        y:
          type: integer
          example: 200
        width:
          type: integer
          example: 300
        height:
          type: integer
          example: 50
...